+++ src/main/java/org/example/cafeflow/cafe/dto/ResponseCafeDto.java
@@ -14,7 +14,7 @@ public class ResponseCafeDto {
     private String name;
     private String address;
     private int reviewCount;
-    private Long reviewsRating;
+    private double reviewsRating;
     private String description;
     private String region;
     private LocalDateTime createdAt;
